refactor(account)!: move `aecli account spend` to `aecli spend`
BREAKING CHANGE: `aecli account spend` renamed to `aecli spend`
