+++ src/aecli-spend.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+import program from './commands/spend.js';
+import { runProgram } from './utils/CliError.js';
+
+await runProgram(program);
