+++ test/account.js
@@ -74,75 +74,6 @@ Secret Key ______________________________ ${keypair.secretKey}
       .to.be.a('string');
   });
 
-  it('Spend coins to another wallet', async () => {
-    const amount = 100;
-    const { publicKey } = generateKeyPair();
-    const resJson = await executeAccount([
-      'spend', WALLET_NAME, '--password', 'test', publicKey, amount, '--json',
-    ]);
-    const receiverBalance = await sdk.getBalance(publicKey);
-    expect(+receiverBalance).to.be.equal(amount);
-
-    expect(resJson.tx.fee).to.be.a('string');
-    expect(resJson).to.eql({
-      blockHash: resJson.blockHash,
-      blockHeight: resJson.blockHeight,
-      encodedTx: resJson.encodedTx,
-      hash: resJson.hash,
-      rawTx: resJson.rawTx,
-      signatures: [resJson.signatures[0]],
-      tx: {
-        amount: '100',
-        fee: resJson.tx.fee,
-        nonce: 1,
-        payload: 'ba_Xfbg4g==',
-        recipientId: resJson.tx.recipientId,
-        senderId: resJson.tx.senderId,
-        ttl: resJson.tx.ttl,
-        type: 'SpendTx',
-        version: 1,
-      },
-    });
-
-    const res = await executeAccount([
-      'spend', WALLET_NAME, '--password', 'test', publicKey, amount,
-    ]);
-    const lineEndings = res.split('\n').map((l) => l.split(' ').at(-1));
-    expect(res).to.be.equal(`
-Transaction mined
-Tx hash _________________________________ ${lineEndings[1]}
-Block hash ______________________________ ${lineEndings[2]}
-Block height ____________________________ ${lineEndings[3]}
-Signatures ______________________________ ${lineEndings[4]}
-Tx Type _________________________________ SpendTx
-Sender account __________________________ ${resJson.tx.senderId}
-Recipient account _______________________ ${resJson.tx.recipientId}
-Amount __________________________________ 100
-Payload _________________________________ ba_Xfbg4g==
-Fee _____________________________________ ${resJson.tx.fee}
-Nonce ___________________________________ 2
-TTL _____________________________________ ${lineEndings[12]}
-Version _________________________________ 1
-    `.trim());
-  });
-
-  it('Spend coins to another wallet in ae', async () => {
-    const receiverKeys = generateKeyPair();
-    const { tx: { fee } } = await executeAccount([
-      'spend', WALLET_NAME, '--password', 'test', '--json',
-      receiverKeys.publicKey, '1ae', '--fee', '0.02ae',
-    ]);
-    expect(await sdk.getBalance(receiverKeys.publicKey)).to.be.equal('1000000000000000000');
-    expect(fee).to.be.equal('20000000000000000');
-  });
-
-  it('Spend percent of coins to account', async () => {
-    const { publicKey } = generateKeyPair();
-    const balanceBefore = await sdk.getBalance(sdk.address);
-    await executeAccount(['spend', WALLET_NAME, '--password', 'test', publicKey, '42%']);
-    expect(+await sdk.getBalance(publicKey)).to.be.equal(balanceBefore * 0.42);
-  });
-
   it('Sign message', async () => {
     const data = 'Hello world';
     const signedMessage = await executeAccount(['sign-message', WALLET_NAME, data, '--json', '--password', 'test']);
