+++ test/index.js
@@ -84,9 +84,12 @@ export async function executeProgram(program, args) {
   try {
     const allArgs = [
       ...args.map((arg) => arg.toString()),
-      ...[
-        'config', 'select-node', 'select-compiler',
-      ].includes(args[0]) ? [] : ['--url', url],
+      ...['config', 'select-node', 'select-compiler'].includes(args[0])
+      || (
+        // eslint-disable-next-line no-underscore-dangle
+        program._name === 'aecli account'
+        && ['save', 'create', 'address', 'sign-message', 'verify-message'].includes(args[0])
+      ) ? [] : ['--url', url],
       ...[
         'compile', 'deploy', 'call', 'encode-calldata', 'decode-call-result',
       ].includes(args[0]) && !args.includes('--compilerUrl') ? ['--compilerUrl', compilerUrl] : [],
